wrote and passed 1st test for rating submission, reminder about query and format differences

+++ db/manage-tables.js
@@ -19,7 +19,7 @@ const createTables = () => {
     upvotes INT DEFAULT 0,
     downvotes INT DEFAULT 0,
     rating_count INT DEFAULT 0 NOT NULL,
-    rating_sum INT DEFAULT 0 NOT NULL,
+    rating_sum DECIMAL DEFAULT 0 NOT NULL,
     rating DECIMAL DEFAULT 1.0 NOT NULL
 );`);
     });
